+++ zhmccli/_cmd_lpar.py
@@ -28,7 +28,7 @@ from ._helper import print_properties, print_resources, abort_if_false, \
     options_to_properties, original_options, COMMAND_OPTIONS_METAVAR, \
     part_console, click_exception, add_options, LIST_OPTIONS, TABLE_FORMATS, \
     hide_property, ASYNC_TIMEOUT_OPTIONS, API_VERSION_HMC_2_14_0, \
-    absolute_capping_value
+    absolute_capping_value, parse_yaml_flow_style
 from ._cmd_cpc import find_cpc
 from ._cmd_certificates import find_certificate
 
@@ -372,9 +372,34 @@ def lpar_show(cmd_ctx, cpc, lpar, **options):
                  'Use of the --zaware-master-pw option is deprecated; use the '
                  '--zaware-master-password option instead.')
 # TODO: Change zAware master password option to ask for password
-# TODO: Support for 'zaware-network-info' property
-# TODO: Support for 'zaware-gateway-info' property
-# TODO: Support for 'zaware-dns-info' property
+@optgroup.option('--zaware-network-info', type=str, required=False,
+                 help='The new list of networks available to IBM zAware, in '
+                 'YAML Flow Collection style. '
+                 'Each list item must be a "zaware-network" object (described '
+                 'in the HMC WS-API book). '
+                 'A minimum of 1 network and a maximum of 100 networks are '
+                 'permitted. The specified list fully replaces the existing '
+                 'list in the HMC. '
+                 'Example: --zaware-network-info "[{port: 444, ipaddr-type: '
+                 'static, vlan-id: 53, static-ip-info: {type: ipv4, '
+                 'ip-address: \'10.11.12.13\', prefix: 24}}]"')
+@optgroup.option('--zaware-gateway-info', type=str, required=False,
+                 help='The new default gateway IP address information for IBM '
+                 'zAware, as an "ip-info" object (described in the HMC WS-API '
+                 'book) in YAML Flow Collection style. '
+                 'An empty string removes the default gateway IP address. '
+                 'Example: --zaware-gateway-info "{type: ipv4, ip-address: '
+                 '\'10.11.12.13\', prefix: 24}"')
+@optgroup.option('--zaware-dns-info', type=str, required=False,
+                 help='The new list of DNS IP addresses for IBM zAware, in '
+                 'YAML Flow Collection style. '
+                 'Each list item must be an "ip-info" object (described in the '
+                 'HMC WS-API book). '
+                 'A minimum of 0 addresses and a maximum of 2 addresses are '
+                 'permitted. The specified list fully replaces the existing '
+                 'list in the HMC. '
+                 'Example: --zaware-dns-info "[{type: ipv4, ip-address: '
+                 '\'10.11.12.13\', prefix: 24}]"')
 @optgroup.group('SSC configuration (only applicable to SSC LPARs and only '
                 'supported on z13)')
 @optgroup.option('--ssc-host-name', type=str, required=False,
@@ -391,9 +416,34 @@ def lpar_show(cmd_ctx, cpc, lpar, **options):
                  'Use of the --ssc-master-pw option is deprecated; use the '
                  '--ssc-master-password option instead.')
 # TODO: Change SSC master password option to ask for password
-# TODO: Support for 'ssc-network-info' property
-# TODO: Support for 'ssc-gateway-info' property
-# TODO: Support for 'ssc-dns-info' property
+@optgroup.option('--ssc-network-info', type=str, required=False,
+                 help='The new list of networks available to the SSC '
+                 'appliance, in YAML Flow Collection style. '
+                 'Each list item must be a "ssc-network" object (described '
+                 'in the HMC WS-API book). '
+                 'A minimum of 1 network and a maximum of 100 networks are '
+                 'permitted. The specified list fully replaces the existing '
+                 'list in the HMC. '
+                 'Example: --ssc-network-info "[{port: 444, ipaddr-type: '
+                 'static, vlan-id: 53, static-ip-info: {type: ipv4, '
+                 'ip-address: \'10.11.12.13\', prefix: 24}}]"')
+@optgroup.option('--ssc-gateway-info', type=str, required=False,
+                 help='The new default gateway IP address information for the '
+                 'SSC appliance, as an "ip-info" object (described in the '
+                 'HMC WS-API book) in YAML Flow Collection style. '
+                 'An empty string removes the default gateway IP address. '
+                 'Example: --ssc-gateway-info "{type: ipv4, ip-address: '
+                 '\'10.11.12.13\', prefix: 24}"')
+@optgroup.option('--ssc-dns-info', type=str, required=False,
+                 help='The new list of DNS IP addresses for the SSC appliance, '
+                 'in YAML Flow Collection style. '
+                 'Each list item must be an "ip-info" object (described in the '
+                 'HMC WS-API book). '
+                 'A minimum of 0 addresses and a maximum of 2 addresses are '
+                 'permitted. The specified list fully replaces the existing '
+                 'list in the HMC. '
+                 'Example: --ssc-dns-info "[{type: ipv4, ip-address: '
+                 '\'10.11.12.13\', prefix: 24}]"')
 @click.pass_obj
 def lpar_update(cmd_ctx, cpc, lpar, **options):
     """
@@ -401,6 +451,11 @@ def lpar_update(cmd_ctx, cpc, lpar, **options):
 
     Only the properties will be changed for which a corresponding option is
     specified, so the default for all options is not to change properties.
+
+    Some of the options of this command are documented to be specified in YAML
+    Flow Collection style. That is a format for specifying complex values
+    that are lists or dictionaries as a relatively simple string. The format
+    is described for example in https://www.yaml.info/learn/flowstyle.html.
 
     In addition to the command-specific options shown in this help text, the
     general options (see 'zhmc --help') can also be specified right after the
@@ -883,6 +938,22 @@ def handle_special_lpar_options(cmd_ctx, org_options, properties):
         properties['zaware-master-userid'] = None
     if org_options['zaware-master-password'] == '':
         properties['zaware-master-pw'] = None
+    if org_options['zaware-network-info']:
+        value = parse_yaml_flow_style(
+            cmd_ctx, '--zaware-network-info',
+            org_options['zaware-network-info'])
+        properties['zaware-network-info'] = value
+    if org_options['zaware-gateway-info'] == '':
+        properties['zaware-gateway-info'] = None
+    elif org_options['zaware-gateway-info']:
+        value = parse_yaml_flow_style(
+            cmd_ctx, '--zaware-gateway-info',
+            org_options['zaware-gateway-info'])
+        properties['zaware-gateway-info'] = value
+    if org_options['zaware-dns-info']:
+        value = parse_yaml_flow_style(
+            cmd_ctx, '--zaware-dns-info', org_options['zaware-dns-info'])
+        properties['zaware-dns-info'] = value
 
     if org_options['ssc-host-name'] == '':
         properties['ssc-host-name'] = None
@@ -890,6 +961,20 @@ def handle_special_lpar_options(cmd_ctx, org_options, properties):
         properties['ssc-master-userid'] = None
     if org_options['ssc-master-password'] == '':
         properties['ssc-master-pw'] = None
+    if org_options['ssc-network-info']:
+        value = parse_yaml_flow_style(
+            cmd_ctx, '--ssc-network-info', org_options['ssc-network-info'])
+        properties['ssc-network-info'] = value
+    if org_options['ssc-gateway-info'] == '':
+        properties['ssc-gateway-info'] = None
+    elif org_options['ssc-gateway-info']:
+        value = parse_yaml_flow_style(
+            cmd_ctx, '--ssc-gateway-info', org_options['ssc-gateway-info'])
+        properties['ssc-gateway-info'] = value
+    if org_options['ssc-dns-info']:
+        value = parse_yaml_flow_style(
+            cmd_ctx, '--ssc-dns-info', org_options['ssc-dns-info'])
+        properties['ssc-dns-info'] = value
 
 
 def cmd_lpar_update(cmd_ctx, cpc_name, lpar_name, options):
@@ -946,7 +1031,14 @@ def cmd_lpar_update(cmd_ctx, cpc_name, lpar_name, options):
         'absolute-icf-capping': None,
 
         'zaware-master-password': 'zaware-master-pw',
+        'zaware-network-info': None,
+        'zaware-gateway-info': None,
+        'zaware-dns-info': None,
+
         'ssc-master-password': 'ssc-master-pw',
+        'ssc-network-info': None,
+        'ssc-gateway-info': None,
+        'ssc-dns-info': None,
     }
     org_options = original_options(options)
     properties = options_to_properties(org_options, name_map)